project/neta/pyproject.toml
[tool.poetry]
name = "neta"
version = "0.1.0"
description = "Automate WhatsApp messaging with AI chatbots using Playwright"
authors = ["Leonardo Fukui <leofukui@hotmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.0"
playwright = "^1.50.0"
pyperclip = "^1.9.0"
python-dotenv = "^1.0.1"
pywhatkit = "^5.4"
selenium = "^4.31.0"
webdriver-manager = "^4.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.0"
pytest = "^8.3.2"
pytest-playwright = "^0.5.2"

[tool.poetry.scripts]
start = "python main.py"
format = "black ."
lint = "ruff check ."
test = "pytest"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "N", "B"]
ignore = ["E501"]